Update to version 0.5.1
Added better documentation.
Changed a bit of the teleport prompt so as to show just the name and not the stringed version of the player object.
Updated "sqltest" to "playerupdate" and made it check whether each found player is already in the database or not, and it will store them if they are not.